+++ src/bda/plone/shop/shipping.py
@@ -5,6 +5,7 @@ from bda.plone.shipping.interfaces import IShippingSettings
 from bda.plone.shipping.interfaces import IShippingItem
 from bda.plone.shop import message_factory as _
 from bda.plone.shop.cartdata import CartItemCalculator
+from bda.plone.shop.utils import format_amount
 from bda.plone.shop.utils import get_shop_settings
 from bda.plone.shop.utils import get_shop_shipping_settings
 from decimal import Decimal
@@ -49,13 +50,8 @@ class DefaultShipping(Shipping):
         flat_shipping_cost = Decimal(str(settings.flat_shipping_cost))
         item_shipping_cost = Decimal(str(settings.item_shipping_cost))
         shipping_vat = Decimal(str(settings.shipping_vat))
-        def quantize(val):
-            val = val.quantize(Decimal('1.00'))
-            if bool(val % 2):
-                return str(val).replace('.', ',')
-            return str(val.quantize(Decimal('1'))) + ',-'
         def gross(val):
-            return quantize(val + (val / Decimal(100) * shipping_vat))
+            return format_amount(val + (val / Decimal(100) * shipping_vat))
         # no shipping costs
         if not flat_shipping_cost and not item_shipping_cost:
             return _(u"free_shipping", default=u"Free Shipping")
@@ -107,7 +103,7 @@ class DefaultShipping(Shipping):
             return Message(msg, mapping={
                 'flat': gross(flat_shipping_cost),
                 'item': gross(item_shipping_cost),
-                'limit': quantize(free_shipping_limit),
+                'limit': format_amount(free_shipping_limit),
                 'currency': currency,
             })
         # flat costs only
@@ -126,7 +122,7 @@ class DefaultShipping(Shipping):
                                 u"${limit} ${currency}")
             return Message(msg, mapping={
                 'flat': gross(flat_shipping_cost),
-                'limit': quantize(free_shipping_limit),
+                'limit': format_amount(free_shipping_limit),
                 'currency': currency,
             })
         # item costs only
@@ -145,7 +141,7 @@ class DefaultShipping(Shipping):
                                 u"price above ${limit} ${currency}")
             return Message(msg, mapping={
                 'item': gross(item_shipping_cost),
-                'limit': quantize(free_shipping_limit),
+                'limit': format_amount(free_shipping_limit),
                 'currency': currency,
             })