+++ src/bda/plone/shop/utils.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from zope.component import getUtility
 from plone.registry.interfaces import IRegistry
 from bda.plone.shop.interfaces import IShopArticleSettings
@@ -7,6 +8,13 @@ from bda.plone.shop.interfaces import IShopShippingSettings
 from bda.plone.shop.interfaces import IShopTaxSettings
 from bda.plone.shop.interfaces import INotificationTextSettings
 from bda.plone.shop.interfaces import IPaymentTextSettings
+
+
+def format_amount(val):
+    val = val.quantize(Decimal('1.00'))
+    if bool(val % 2):
+        return str(val).replace('.', ',')
+    return str(val.quantize(Decimal('1'))) + ',-'
 
 
 def get_shop_settings():